+++ client/CHANGELOG.md
@@ -48,6 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `EventPayload` renamed to `ParticipationEventPayload`;
 - `Client::{subscriber(), subscribe(), unsubscribe()}` take now `&self` instead of `&mut self`;
 - `{MqttManager, MqttTopicManager}::new()` take now `&'a Client` instead of `&'a mut Client`;
+- `Error::Pow` to `Error::NonceNotFound`;
+- `finish_nonce` takes a `F: Fn(&[u8]) -> Option<u64>` instead of a `F: Fn(&[u8]) -> Result<u64, PowError>`;
 
 ### Removed